refactor: 컨벤션 적용, 피드백 적용
